requirements.txt
botocore==1.10.56
colorama==0.3.9
diff-match-patch==20121119
Django==1.11.13
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.7
docutils==0.14
et-xmlfile==1.0.1
future==0.16.0
futures==3.2.0
hdfs3==0.3.0
httplib2==0.11.3
jdcal==1.4
jmespath==0.9.3
MySQL-python==1.2.5
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.1.0
pyasn1==0.4.3
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
